requirements.txt
astroid==2.2.5
certifi==2019.6.16
chardet==3.0.4
colorama==0.4.1
defusedxml==0.6.0
Django==2.2.4
django-allauth==0.39.1
django-crispy-forms==1.7.2
django-rest-auth==0.9.5
djangorestframework==3.10.2
html5lib==1.0.1
idna==2.8
isort==4.3.21
lazy-object-proxy==1.4.2
mccabe==0.6.1
oauthlib==3.1.0
Pillow==6.1.0
pylint==2.3.1
PyPDF2==1.26.0
python3-openid==3.1.0
pytz==2019.2
reportlab==3.5.23
requests==2.22.0
requests-oauthlib==1.2.0
six==1.12.0
sqlparse==0.3.0
typed-ast==1.4.0
urllib3==1.25.3
webencodings==0.5.1
wrapt==1.11.2
xhtml2pdf==0.2.3
astroid==2.2.5
certifi==2019.6.16
chardet==3.0.4
colorama==0.4.1
defusedxml==0.6.0
Django==2.2.4
django-allauth==0.39.1
django-crispy-forms==1.7.2
django-rest-auth==0.9.5
djangorestframework==3.10.2
html5lib==1.0.1
idna==2.8
isort==4.3.21
lazy-object-proxy==1.4.2
mccabe==0.6.1
oauthlib==3.1.0
Pillow==6.1.0
pylint==2.3.1
PyPDF2==1.26.0
python3-openid==3.1.0
pytz==2019.2
reportlab==3.5.23
requests==2.22.0
requests-oauthlib==1.2.0
six==1.12.0
sqlparse==0.3.0
typed-ast==1.4.0
urllib3==1.25.3
webencodings==0.5.1
wrapt==1.11.2
xhtml2pdf==0.2.3
astroid==2.2.5
certifi==2019.6.16
chardet==3.0.4
colorama==0.4.1
defusedxml==0.6.0
Django==2.2.4
django-allauth==0.39.1
django-crispy-forms==1.7.2
django-rest-auth==0.9.5
djangorestframework==3.10.2
html5lib==1.0.1
idna==2.8
isort==4.3.21
lazy-object-proxy==1.4.2
mccabe==0.6.1
oauthlib==3.1.0
Pillow==6.1.0
pylint==2.3.1
PyPDF2==1.26.0
python3-openid==3.1.0
pytz==2019.2
reportlab==3.5.23
requests==2.22.0
requests-oauthlib==1.2.0
six==1.12.0
sqlparse==0.3.0
typed-ast==1.4.0
urllib3==1.25.3
webencodings==0.5.1
wrapt==1.11.2
xhtml2pdf==0.2.3
